chores: - divide uiManager into 4 different parts - ui_manager: manages application views - ui_menu: code related to menus - ui_sequencer: code realted to sequencer - ui_common: common code & helper functions - rename uiManger to ui_manager
